chore: added preliminary internal unit tests
