+++ .vscode/launch.json
@@ -1,0 +1,16 @@
+{
+  "configurations": [
+    {
+      "args": ["run", "${relativeFile}"],
+      "autoAttachChildProcesses": true,
+      "console": "integratedTerminal",
+      "name": "Debug Current Test File",
+      "program": "${workspaceRoot}/node_modules/vitest/vitest.mjs",
+      "request": "launch",
+      "skipFiles": ["<node_internals>/**", "**/node_modules/**"],
+      "smartStep": true,
+      "type": "node"
+    }
+  ],
+  "version": "0.2.0"
+}
