+++ package.json
@@ -14,7 +14,8 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
     "pretty-quick": "^3.1.3",
-    "typescript": "^4.7.4"
+    "typescript": "^4.7.4",
+    "vitest": "^0.31.0"
   },
   "main": "src/index.js",
   "husky": {
@@ -43,7 +44,8 @@
     "lint": "yarn eslint \"src/*.ts\"  --max-warnings 0 --report-unused-disable-directives",
     "lint:md": "yarn markdownlint \"**/*.md\" \".github/**/*.md\"",
     "lint:spelling": "cspell \"**\" \".github/**/*\"",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "test": "vitest"
   },
   "version": "0.2.1"
 }
