+++ vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    clearMocks: true,
+    exclude: ["lib", "node_modules", "src/**/*.js", "src/*/*.js"],
+  },
+});
